skills/gitlab/requirements.txt
certifi==2022.6.15;
charset-normalizer==2.1.0;
idna==2.10;
python-gitlab==2.9.0;
pyyaml==6.0;
requests-toolbelt==0.9.1;
requests==2.23;
urllib3==1.25.11;
chardet==3.0.4;
